chore: refactor-union-file
Refactor of the union.ts file to a single /types file. removed union
export.
