+++ packages/davinci-client/src/lib/client.store.utils.ts
@@ -6,13 +6,12 @@
  */
 import { configureStore } from '@reduxjs/toolkit';
 
-import type { RequestMiddleware } from '@forgerock/effects/types';
+import type { ActionTypes, RequestMiddleware } from '@forgerock/effects/types';
 import { configSlice } from './config.slice.js';
 import { nodeSlice } from './node.slice.js';
 import { davinciApi } from './davinci.api.js';
 import { ErrorNode, ContinueNode, StartNode, SuccessNode } from '../types.js';
 import { wellknownApi } from './wellknown.api.js';
-import { ActionTypes } from '@forgerock/effects/unions';
 
 export function createClientStore<ActionType extends ActionTypes>({
   requestMiddleware,
