+++ packages/effects/package.json
@@ -11,8 +11,7 @@
       "default": "./dist/src/index.js"
     },
     "./package.json": "./package.json",
-    "./types": "./dist/src/lib/request.types.js",
-    "./unions": "./dist/src/lib/request.unions.js"
+    "./types": "./dist/src/lib/request.types.js"
   },
   "main": "./dist/src/index.js",
   "module": "./dist/src/index.js",
