+++ packages/effects/src/lib/request.types.ts
@@ -11,8 +11,18 @@ import {
   QueryReturnValue,
 } from '@reduxjs/toolkit/query';
 
-import type { ActionTypes } from './request.unions.js';
+export const actionTypes = {
+  start: 'DAVINCI_START',
+  next: 'DAVINCI_NEXT',
+  flow: 'DAVINCI_FLOW',
+  success: 'DAVINCI_SUCCESS',
+  error: 'DAVINCI_ERROR',
+  failure: 'DAVINCI_FAILURE',
+  resume: 'DAVINCI_RESUME',
+} as const;
 
+export type ActionTypes = (typeof actionTypes)[keyof typeof actionTypes];
+export type EndpointTypes = keyof typeof actionTypes;
 export interface Action<Type extends ActionTypes = ActionTypes, Payload = unknown> {
   type: Type;
   payload: Payload;
